Changed all local variable references to getters and setters. Changed all variables to private.

+++ src/TradeScreen/TradeScreen.java
@@ -6,8 +6,8 @@ import OrderManager.Order;
 
 public interface TradeScreen {
 	public enum api{newOrder,price,fill,cross};
-	public void newOrder(int id,Order order) throws IOException, InterruptedException;
-	public void acceptOrder(int id) throws IOException;
-	public void sliceOrder(int id,int sliceSize) throws IOException;
-	public void price(int id,Order o) throws InterruptedException, IOException;
+	public void newOrder(long id,Order order) throws IOException, InterruptedException;
+	public void acceptOrder(long id) throws IOException;
+	public void sliceOrder(long id,long sliceSize) throws IOException;
+	public void price(long id,Order o) throws InterruptedException, IOException;
 }
